Add project to read CLZ XML data with `quick-xml`
So far we can read all the tags and count the "book" tags.

Next we will start loading those into structures and try to deal with
the various kinds of data -- e.g., text, numbers, dates/timestamps,
etc. Then we'll try to do something with the data, like shoving it
into a database.