Extract book authors in parser and insert them into database
Finishes updating the parse and adds database code to insert authors
and `author_book` rows connecting authors to books.

TODO: There is no attempt here at disambiguation or duplicate detection.
Our idea is to do this in a later step, but it might be desirable to
add code to do this as the XML file is parsed at some point. Maybe for
that we can at least add some code to filter out new records that are
identical to ones in the database already.