Report db update results to main thread and wait for all task to complete
This has the database thread report the result of its insert operations
for book and author data back to the main thread. We've create a task id
for each piece of book data the parser thread finds and we use this to
create a set of database tasks that the main thread waits for the
database thread to execute before the program is allowed to complete.