Merge in modified Deno + Vite + React server and client code
The original was here: https://github.com/denoland/react-vite-ts-template

We have Dockerized it and modified it to have a React Table component
that will contain data on books from our collection. We need to add
an endpoint that serves actual book data from the database for the
React frontend to use.