TUI: Implement parser interface with mock parser function
We launch a goroutine that calls the parser and sends book records
to a channel until it is complete, at which point it sends "Done".
Our second, `DataImportModel` Bubbletea model creates a `tea.Cmd`
that blocks on receiving from this channel until it receives a value
or the channel closes. When it receives a value it returns a `tea.Msg`
with the `BookRecord` value, or the string "Done" if the parser is
done.

TODO: This version has a mock parser that just generates a series of
sample book titles, but we can now connect it to a Golang client for
our Rust gRPC server that hosts the parser endpoint.

TODO: Once that step is done, we will by some means implement inserting
a record into the database when the user chooses to accept that record.
This could be another gRPC endpoint, or it could be an endpoint in our
Go database utility.

One way or another, it has been a useful exercise to design and
implement this feature, because the ideas are useful in many places.