Add copy of .proto file from `parser_grpc`
This will allow us to generate the gRPC client for this service in
Golang, which we'll call from the TUI. It might be worth thinking
about how to handle the duplication. We could write a build script
that copies the file from the Rust directory before buildling, so
we don't have to worry about keeping the files in sync.