Add page selector component to BookTable; add ids to table headers
It might be reasonable to put the page selector a level above, so it
is a sibling of the book table, but this makes the selector part of
the `BookTable` component, beside the React Table.

We add ids to table `th` elements so we can style the columns with
fixed widths, etc.