Call database thread `async fn database_main` in top-level `block_on`
So we create a tokio runtime when we spawn the database thread and
have it run the database main function with `block_on`. This means
the database thread now runs as if it were its own async program that
can communicate with the synchronous main thread by message passing.

This is how we should have done things from the start, but then we
were just making things work ad hoc, and less concerned about making
them efficient.

TODO: If we can get the parser thread running async -- if quick_xml
supports that, specifically -- then we can similarly run the parser
as its own async task.