Parser gRPC: sending all errors back to caller by channel
This replaces log.Fatal calls with reporting back to caller and then
terminating the goroutine.

NOTE: The Go specification for `close` says:

"After calling close, and after any previously sent values have been
received, receive operations will return the zero value for the
channel's type without blocking. The multi-valued receive operation
returns a received value along with an indication of whether the
channel is closed."

This is significant because error handling could block on sending to
the channel, as we're sharing the data out channel for error handling,
resulting in a delayed notification that an error has happened.
Fortunately, it's only really an issue in the receive loop, after setup.